SEG-11284: Crashed membase on new 1.8 membase master
When a key maps to the value -2^31 during hashing, std::abs is not
enough to ensure positive value. This fix handles this problem for keys
that produce this value (eg: 36371502474###HARVEST_PROD:ACCOUNTSTATUS).
This problem is visible only when db_shards is not a power of 2.

+++ sqlite-strategies.hh
@@ -115,7 +115,8 @@ protected:
         for(i=0; str[i] != 0x00; i++) {
             h = ((h << 5) + h) ^ str[i];
         }
-        return std::abs(h) % (int)shardCount;
+        // Cast to unsigned to handle the value -2^31
+        return (unsigned int)std::abs(h) % (int)shardCount;
     }
 
     virtual void initDB() {
